Features

+++ src/main/java/com/sasagui/onmangequoi/dish/DishScorer.java
@@ -3,5 +3,5 @@ package com.sasagui.onmangequoi.dish;
 @FunctionalInterface
 public interface DishScorer {
 
-    int score(DishScoringContext context);
+    float score(DishScoringContext context);
 }
